requirements.txt
flask ~= 3.1.0
ollama ~= 0.4.4
python-dotenv ~= 1.0.1
flask-cors ~= 5.0.0